PRSD-1393: landlord record content sweep (#695)
