+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/viewModels/summaryModels/RegisteredPropertylViewModels.kt
@@ -7,7 +7,7 @@ import uk.gov.communities.prsdb.webapp.database.entity.PropertyOwnership
 import uk.gov.communities.prsdb.webapp.helpers.converters.MessageKeyConverter
 import uk.gov.communities.prsdb.webapp.models.dataModels.RegistrationNumberDataModel
 
-data class RegisteredPropertyViewModel(
+data class RegisteredPropertyLocalCouncilViewModel(
     val address: String,
     val registrationNumber: String,
     val localAuthorityName: String,
@@ -18,10 +18,9 @@ data class RegisteredPropertyViewModel(
     companion object {
         fun fromPropertyOwnership(
             propertyOwnership: PropertyOwnership,
-            isLaView: Boolean = false,
             currentUrlKey: Int? = null,
-        ): RegisteredPropertyViewModel =
-            RegisteredPropertyViewModel(
+        ): RegisteredPropertyLocalCouncilViewModel =
+            RegisteredPropertyLocalCouncilViewModel(
                 address = propertyOwnership.property.address.singleLineAddress,
                 registrationNumber =
                     RegistrationNumberDataModel
@@ -38,7 +37,32 @@ data class RegisteredPropertyViewModel(
                 isTenantedMessageKey = MessageKeyConverter.convert(propertyOwnership.isOccupied),
                 recordLink =
                     PropertyDetailsController
-                        .getPropertyDetailsPath(propertyOwnership.id, isLaView)
+                        .getPropertyDetailsPath(propertyOwnership.id, isLaView = true)
+                        .overrideBackLinkForUrl(currentUrlKey),
+            )
+    }
+}
+
+data class RegisteredPropertyLandlordViewModel(
+    val address: String,
+    val registrationNumber: String,
+    val recordLink: String,
+) {
+    companion object {
+        fun fromPropertyOwnership(
+            propertyOwnership: PropertyOwnership,
+            currentUrlKey: Int? = null,
+        ): RegisteredPropertyLandlordViewModel =
+            RegisteredPropertyLandlordViewModel(
+                address = propertyOwnership.property.address.singleLineAddress,
+                registrationNumber =
+                    RegistrationNumberDataModel
+                        .fromRegistrationNumber(
+                            propertyOwnership.registrationNumber,
+                        ).toString(),
+                recordLink =
+                    PropertyDetailsController
+                        .getPropertyDetailsPath(propertyOwnership.id, isLaView = false)
                         .overrideBackLinkForUrl(currentUrlKey),
             )
     }
