+++ src/main/resources/messages.properties
@@ -765,9 +765,9 @@ propertySearch.details.body=If you can't find a property record, the property ma
 
 landlordDetails.title=Landlord record
 landlordDetails.heading=Landlord record
-landlordDetails.removeRecord=Delete record
+landlordDetails.removeRecord=Delete account
 landlordDetails.personalDetails.heading=Personal details
-landlordDetails.personalDetails.lrn=Landlord registration number
+landlordDetails.personalDetails.lrn=Landlord Registration Number
 landlordDetails.personalDetails.registrationDate=Database registration date
 landlordDetails.personalDetails.name=Name
 landlordDetails.personalDetails.dateOfBirth=Date of birth
@@ -781,8 +781,8 @@ landlordDetails.personalDetails.nonEnglandOrWalesAddress=Address (outside UK)
 landlordDetails.personalDetails.englandOrWalesAddress=UK contact address
 landlordDetails.registeredProperties.heading=Registered properties
 landlordDetails.registeredProperties.table.addressHeading=Property address
+landlordDetails.registeredProperties.table.prnHeading=Property Registration Number
 landlordDetails.registeredProperties.table.localAuthorityHeading=Local council
-landlordDetails.registeredProperties.table.licenceHeading=Property licence
 landlordDetails.registeredProperties.table.tenantedHeading=Tenanted
 landlordDetails.registeredProperties.table.registrationNumberHeading=Registration number
 landlordDetails.registeredProperties.table.licensingTypeHeading=Licensing type
