+++ src/main/resources/templates/landlordDetailsView.html
@@ -21,20 +21,16 @@
 
                 <div id="registered-properties-panel-content" th:replace="~{fragments/tabs/tabsPanel :: tabsPanel(${registeredPropertiesTabId}, ~{::#registered-properties-panel-content/content()})}">
                     <table th:replace="~{fragments/tables/table :: table(
-                        ${ { 'landlordDetails.registeredProperties.table.addressHeading', 'landlordDetails.registeredProperties.table.localAuthorityHeading', 'landlordDetails.registeredProperties.table.licenceHeading', 'landlordDetails.registeredProperties.table.tenantedHeading' } },
+                        ${ { 'landlordDetails.registeredProperties.table.addressHeading', 'landlordDetails.registeredProperties.table.prnHeading' } },
                         ~{::tbody},
                         null
                     )}">
                         <tbody class="govuk-table__body">
                             <tr class="govuk_table__row" th:each="property: ${registeredPropertiesList}">
-                                <td class="govuk-table__cell govuk-!-width-one-quarter">
+                                <td class="govuk-table__cell govuk-!-width-one-half">
                                     <a class="govuk-link" th:text="${property.address}" th:href="@{${property.recordLink}}">${property.address}</a>
-                                    <br>
-                                    <span th:remove="tag" th:text="${property.registrationNumber}"></span>
                                 </td>
-                                <td class="govuk-table__cell govuk-!-width-one-quarter" th:text="${property.localAuthorityName}"></td>
-                                <td class="govuk-table__cell govuk-!-width-one-quarter" th:text="#{${property.licenseTypeMessageKey}}"></td>
-                                <td class="govuk-table__cell govuk-!-width-one-quarter" th:text="#{${property.isTenantedMessageKey}}"></td>
+                                <td class="govuk-table__cell govuk-!-width-one-half" th:text="${property.registrationNumber}"></td>
                             </tr>
                         </tbody>
                     </table>
